[lint package refactor] Create a file for the Run class
